update my feed folder
