+++ micro-service/my-feed-service/go.mod
@@ -3,7 +3,7 @@ module github.com/YukiOnishi1129/techpicks/micro-service/my-feed-service
 go 1.23.4
 
 require (
-	github.com/YukiOnishi1129/checkpicks-protocol-buffers/checkpicks-rpc-go v1.1.0
+	github.com/YukiOnishi1129/checkpicks-protocol-buffers/checkpicks-rpc-go v1.2.3-0.20241231130221-a5f27fd98c60
 	github.com/friendsofgo/errors v0.9.2
 	github.com/joho/godotenv v1.5.1
 	github.com/lib/pq v1.10.9
